Added Contest,Editorial,Subscribers,Webinar

+++ controller/SecretCode.js
@@ -4,32 +4,61 @@ const crypto = require("crypto");
 const User = require("../models/User");
 const SecretCode = require("../models/SecretCode");
 
-const email = require("./email");
+const Email = require("../helper/email");
 const generatecode = () => {
-    return crypto.randomBytes(12).toString("hex");
+    return crypto.randomBytes(64).toString("hex");
 };
 
-// url: `${baseUrl}/api/verify/${user._id}/${SecretCode}`;
+exports.codegen = async (email) => {
+    let code = generatecode();
+    try {
+        const user = await User.findOne({ email: email });
+        const redundantCodes = await SecretCode.find({ user_id: user._id });
+        if (redundantCodes.length > 0) {
+            redundantCodes.forEach(async (secretCode) => {
+                delid = secretCode._id;
+                await SecretCode.findByIdAndDelete(delid);
+            });
+        }
+
+        const secretCode = new SecretCode({
+            code: code,
+            email_id: email,
+            user_id: user._id,
+        });
+
+        await secretCode.save();
+
+        let link = `${process.env.BASEURL}:${
+            process.env.PORT || 4000
+        }/api/verify/${user._id}&${code}`;
+
+        await Email.sendLink(email, link);
+        return true;
+    } catch (err) {
+        return false;
+    }
+};
 
 exports.sendcode = async (req, res) => {
     let code = generatecode();
-
     try {
         const user = await User.findOne({ email: req.body.email });
-
         if (!user) {
             return res.status(404).json({
-                error: "User couldn't be found.",
+                error: true,
+                errorBody: "No User Found.",
             });
         }
 
         const redundantCodes = await SecretCode.find({ user_id: user._id });
-
-        redundantCodes.forEach(async (secretCode) => {
-            delid = secretCode._id;
-            console.log(delid);
-            await SecretCode.findByIdAndDelete(delid);
-        });
+        if (redundantCodes.length > 0) {
+            redundantCodes.forEach(async (secretCode) => {
+                delid = secretCode._id;
+                console.log(delid);
+                await SecretCode.findByIdAndDelete(delid);
+            });
+        }
 
         const secretCode = new SecretCode({
             code: code,
@@ -38,22 +67,25 @@ exports.sendcode = async (req, res) => {
         });
 
         await secretCode.save();
-
-        let link = `${process.env.BASEURL}:${process.env.PORT || 4000}${
-            req.baseUrl
-        }/${user._id}&${code}`;
+        let link = `${process.env.BASEURL}:${
+            process.env.PORT || 4000
+        }/api/verify/${user._id}&${code}`;
 
         try {
-            await email.sendLink(req.body.email, link);
+            await Email.sendLink(req.body.email, link);
         } catch (err) {
             console.log(err);
         }
 
         res.status(200).json({
-            msg: `${req.body.email} has been sent code.`,
+            error: false,
+            message: "Mail Sent Succefully",
         });
     } catch (err) {
-        console.log(err);
+        return res.status(500).json({
+            error: true,
+            errorBody: `Internal Server Error.`,
+        });
     }
 };
 
@@ -66,7 +98,8 @@ exports.verifycode = async (req, res) => {
 
         if (!user) {
             return res.status(404).json({
-                error: "User couldn't be found.",
+                error: true,
+                errorBody: "No User Found.",
             });
         }
 
@@ -76,7 +109,8 @@ exports.verifycode = async (req, res) => {
 
         if (!secretCode) {
             return res.status(404).json({
-                error: "User Code Expired, Please Generate new one",
+                error: true,
+                errorBody: "Code Not Found, Generate New One.",
             });
         } else if (secretCode.code === code) {
             await user.updateOne({
@@ -86,10 +120,14 @@ exports.verifycode = async (req, res) => {
             });
 
             res.status(200).json({
-                msg: ` verifed has been updated successfully.`,
+                error: false,
+                message: ` User Account Status has been updated successfully.`,
             });
         }
     } catch (err) {
-        console.log(err);
+        return res.status(500).json({
+            error: true,
+            errorBody: `Internal Server Error.`,
+        });
     }
 };
